+++ src/ambition/7kaaInterface/input.cc
@@ -30,6 +30,7 @@
 #include <SDL.h>
 
 #include "pragma_silence_7kaa_warnings.hh"
+#include "OCONFIG.h"
 #include "KEY.h"
 #include "OBUTT3D.h"
 #include "OFIRMA.h"
@@ -183,6 +184,9 @@ const std::map<Action, KeyEventType> KEY_EVENT_MAP = {
   { Action::Common_Cancel, KEYEVENT_CANCEL },
   { Action::Common_Confirm, KEYEVENT_CONFIRM },
   { Action::Common_Reward, KEYEVENT_UNIT_REWARD },
+
+  { Action::Speed_SetToUnlimited, KEYEVENT_SPEED_SET_TO_UNLIMITED },
+  { Action::Speed_TogglePause, KEYEVENT_SPEED_TOGGLE_PAUSE },
 
   { Action::Spy_Assassinate, KEYEVENT_SPY_ASSASSINATE },
   { Action::Spy_Bribe, KEYEVENT_SPY_BRIBE },
@@ -783,6 +787,46 @@ void detectScenarioScroll(
     },
     []() { return SLOT_COUNT - 1; }
   );
+}
+
+bool detectSpeedChangeKeys(
+  const unsigned int keyCode
+) {
+  if (!Ambition::config.enhancementsAvailable()) {
+    return false;
+  }
+
+  constexpr auto STEP_STANDARD = 3;
+  constexpr auto SPEED_UNLIMITED = 99;
+
+  if (keyCode == getKeyEvent(Action::Speed_SetToUnlimited)) {
+    sys.set_speed(SPEED_UNLIMITED);
+    sys.user_pause_flag = 0;
+    return true;
+  }
+
+  if (keyCode == getKeyEvent(Action::Speed_TogglePause)) {
+    sys.set_speed(0);
+    sys.user_pause_flag = !config.frame_speed;
+    return true;
+  }
+
+  /* 1–8 are already handled by 7kaa.
+   *
+   * These are here after the above bound events, to allow people to rebind `9`
+   * and `0` back to unlimited and pause, if they wish. */
+  if (keyCode == '9') {
+    sys.set_speed(9 * STEP_STANDARD);
+    sys.user_pause_flag = 0;
+    return true;
+  }
+  if (keyCode == '0') {
+    sys.set_speed(10 * STEP_STANDARD);
+    sys.user_pause_flag = 0;
+    return true;
+  }
+
+  return false;
 }
 
 bool detectSpyScroll(